+++ Implementierung/svg.h
@@ -2,9 +2,9 @@
 #include <stdio.h>
 
 void save_as_svg(
-        const uint64_t *x,
-        const uint64_t *y,
-        unsigned int degree,
-        unsigned int scale,
+        const uint32_t *x,
+        const uint32_t *y,
+        uint32_t degree,
+        uint32_t scale,
         FILE *output_file
 );
